상품 목록, 찜 목록, 판매 목록에 상품이 없는 경우 기본 메시지가 출력되도록 설정
