+++ src/main/java/teamproject/usedmarket/config/WebMvcConfig.java
@@ -3,8 +3,11 @@ package teamproject.usedmarket.config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import teamproject.usedmarket.web.interceptor.LogInterceptor;
+import teamproject.usedmarket.web.interceptor.LoginCheckInterceptor;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
@@ -16,6 +19,20 @@ public class WebMvcConfig implements WebMvcConfigurer {
                 .addResourceHandler("/files/**")
 //                .addResourceLocations("file:///C:/Users/82109/Desktop/spring_img/");
                 .addResourceLocations("file:/Users/kimgang/Documents/SpringProject/imageFile/");
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new LogInterceptor())
+                .order(1)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/css/**", "/*.ico", "/error");
+
+        registry.addInterceptor(new LoginCheckInterceptor())
+                .order(2)
+                .addPathPatterns("/**")
+                .excludePathPatterns("/", "/members/add", "/login", "/logout",
+                        "/css/**", "/*.ico", "/error");
     }
 
 }
