--- src/main/resources/templates/entities.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head th:replace="fragments/header :: header"/>
-<body>
-<h2>Entity List</h2>
-
-<div>
-<table border="1" class="table table-striped">
-    <thead>
-    <tr>
-        <th>번호</th>
-        <th>제목</th>
-        <th>등록 일자</th>
-        <th>이미지 보기</th>
-        <th></th>
-        <!-- Add other columns based on your YourEntity class -->
-    </tr>
-    </thead>
-    <tbody>
-    <tr th:each="entity : ${entities}">
-        <td>
-            <a th:href="@{/items/{itemId}(itemId=${entity.itemId})}" th:text="${entity.itemId}"> </a>
-        </td>
-        <td>
-        <a th:href="@{/items/{itemId}(itemId=${entity.itemId})}" th:text="${entity.title}"> </a>
-        </td>
-        <td th:text="${#dates.format(entity.createDatetime, 'yyyy-MM-dd HH:mm:ss')}">등록 일자</td>
-        <td>
-        <block th:each="image : ${images}">
-            <!--                    <img th:if="${item.filename}" th:src="@{|/files/${item.getFilename()}|}" width="100" height="100"/>-->
-            <img th:if="${entity.itemId == image.itemId} and ${image.repImageCheck == true}" th:src="@{|/files/${image.fileName}|}" width="200" height="200"/>
-        </block>
-            </td>
-        <!-- Add other columns based on your YourEntity class -->
-    </tr>
-    </tbody>
-</table>
-</div>
-
-<div th:if="${totalPages > 1}">
-    <ul class="pagination">
-        <li th:class="${currentPage == 1} ? 'page-item disabled' : 'page-item'">
-            <a th:href="@{/entities(page=1)}" class="page-link">First</a>
-        </li>
-        <li th:each="pageNumber : ${#numbers.sequence(1, totalPages)}" th:class="${currentPage == pageNumber} ? 'page-item active' : 'page-item'">
-            <a th:href="@{/entities(page=${pageNumber})}" class="page-link" th:text="${pageNumber}"></a>
-        </li>
-        <li th:class="${currentPage == totalPages} ? 'page-item disabled' : 'page-item'">
-            <a th:href="@{/entities(page=${totalPages})}" class="page-link">Last</a>
-        </li>
-    </ul>
-</div>
-<div th:replace="fragments/footer :: footer"/>
-</body>
-</html>
